refactor: rename url_utils to tenant
